added a check for existing files

+++ build-scripts/build-lammps.sh
@@ -32,6 +32,15 @@ fi
 
 
 # Download and build LAMMPS 
+if [[ -e stable_${LAMMPS_VERSION}.zip ]]; then
+  echo "the stable_${LAMMPS_VERSION}.zip that we want to download already exists, please delete it to be sure that the file is properly downloaded"
+  exit 0
+fi
+if [[ -e lammps-stable_${LAMMPS_VERSION} ]]; then
+  echo "the lammps-stable_${LAMMPS_VERSION} folder from unzipping the stable_${LAMMPS_VERSION}.zip already exists, please delete it to be sure that the build is clean"
+  exit 0
+fi
+
 wget https://github.com/lammps/lammps/archive/stable_${LAMMPS_VERSION}.zip
 unzip stable_${LAMMPS_VERSION}.zip 
 rm -f stable_${LAMMPS_VERSION}.zip
